# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/matching/jobs": {
      "get": {
        "operationId": "MatchingController_getMatchedJobs",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Get jobs matched to user profile (rule-based)",
        "tags": ["matching"]
      }
    },
    "/api/matching/ai-recommendations": {
      "get": {
        "operationId": "MatchingController_getAIRecommendations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get AI-powered job recommendations based on profile",
        "tags": ["matching"]
      }
    },
    "/api/matching/feed": {
      "get": {
        "operationId": "MatchingController_getFeed",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Get personalized job feed (AI + Rule-based)",
        "tags": ["matching"]
      }
    },
    "/api/matching/job/{jobId}/semantic": {
      "get": {
        "operationId": "MatchingController_getSemanticMatch",
        "parameters": [
          { "name": "jobId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get detailed AI semantic matching analysis",
        "tags": ["matching"]
      }
    },
    "/api/matching/job/{jobId}": {
      "get": {
        "operationId": "MatchingController_getJobDetail",
        "parameters": [
          { "name": "jobId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Get detailed rule-based matching analysis",
        "tags": ["matching"]
      }
    },
...
    "/api/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markAsRead",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Notification" } }
            }
          }
        },
        "tags": ["Notifications"]
      }
    },
...
    "/api/notifications/{id}": {
      "delete": {
        "operationId": "NotificationsController_remove",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Notifications"]
      }
    },
...
  },
  "info": {
    "title": "AI Job Application Context API",
    "description": "The AI Job Application API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
...
      "ProfileSkill": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "profileId": { "type": "string" },
          "profile": { "$ref": "#/components/schemas/Profile" },
          "name": { "type": "string" },
          "category": {
            "enum": ["professional", "technical", "interpersonal", "domain", "language", "tool"],
            "type": "string"
          },
          "level": { "enum": ["strong", "used_before", "learning"], "type": "string" },
          "contexts": { "type": "array", "items": { "type": "object" } },
          "evidence": { "type": "array", "items": { "type": "object" } },
          "source": { "enum": ["user", "cv_parse", "ai_suggest"], "type": "string" },
          "confidence": { "type": "number" },
          "lastUsedYear": { "type": "number", "nullable": true },
          "possibleDuplicate": { "type": "boolean" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "profileId",
          "profile",
          "name",
          "category",
          "level",
          "contexts",
          "evidence",
          "source",
          "confidence",
          "lastUsedYear",
          "possibleDuplicate",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProfileExperience": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "profileId": { "type": "string" },
          "profile": { "$ref": "#/components/schemas/Profile" },
          "organization": { "type": "string" },
          "role": { "type": "string" },
          "employmentType": {
            "enum": ["full_time", "part_time", "freelance", "internship"],
            "type": "string"
          },
          "startDate": { "format": "date-time", "type": "string", "nullable": true },
          "endDate": { "format": "date-time", "type": "string", "nullable": true },
          "responsibilities": { "type": "array", "items": { "type": "object" } },
          "scope": {
            "enum": ["individual", "team", "multi_team", "organization"],
            "type": "string"
          },
          "skillsUsed": { "type": "array", "items": { "type": "string" } },
          "source": { "enum": ["user", "cv_parse", "ai_suggest"], "type": "string" },
          "confidence": { "type": "number" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "profileId",
          "profile",
          "organization",
          "role",
          "employmentType",
          "startDate",
          "endDate",
          "responsibilities",
          "scope",
          "skillsUsed",
          "source",
          "confidence",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProfileProject": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "profileId": { "type": "string" },
          "profile": { "$ref": "#/components/schemas/Profile" },
          "name": { "type": "string" },
          "context": {
            "enum": ["academic", "personal", "freelance", "internal", "volunteer"],
            "type": "string"
          },
          "description": { "type": "string", "nullable": true },
          "role": { "type": "string", "nullable": true },
          "skillsUsed": { "type": "array", "items": { "type": "string" } },
          "outcomes": { "type": "array", "items": { "type": "string" } },
          "source": { "enum": ["user", "cv_parse", "ai_suggest"], "type": "string" },
          "confidence": { "type": "number" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "profileId",
          "profile",
          "name",
          "context",
          "description",
          "role",
          "skillsUsed",
          "outcomes",
          "source",
          "confidence",
          "createdAt",
          "updatedAt"
        ]
      },
      "CareerIntent": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "profileId": { "type": "string" },
          "profile": { "$ref": "#/components/schemas/Profile" },
          "applyNowRoles": { "type": "array", "items": { "type": "string" } },
          "targetRoles": { "type": "array", "items": { "type": "string" } },
          "desiredSeniority": {
            "nullable": true,
            "enum": ["entry", "junior", "mid", "senior", "lead", "manager"],
            "type": "string"
          },
          "salaryExpectation": { "type": "object", "nullable": true },
          "companyPreferences": { "type": "array", "items": { "type": "string" } },
          "industries": { "type": "array", "items": { "type": "string" } },
          "avoid": { "type": "object" },
          "source": { "enum": ["user", "cv_parse", "ai_suggest"], "type": "string" },
          "confidence": { "type": "number" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "profileId",
          "profile",
          "applyNowRoles",
          "targetRoles",
          "desiredSeniority",
          "salaryExpectation",
          "companyPreferences",
          "industries",
          "avoid",
          "source",
          "confidence",
          "updatedAt"
        ]
      },
      "WorkPreferences": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "profileId": { "type": "string" },
          "profile": { "$ref": "#/components/schemas/Profile" },
          "locations": { "type": "array", "items": { "type": "string" } },
          "workMode": { "enum": ["remote", "onsite", "hybrid", "flexible"], "type": "string" },
          "workingHours": { "enum": ["fixed", "flexible", "shift"], "type": "string" },
          "languages": { "type": "array", "items": { "type": "string" } },
          "dealBreakers": { "type": "array", "items": { "type": "string" } },
          "source": { "enum": ["user", "cv_parse", "ai_suggest"], "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "profileId",
          "profile",
          "locations",
          "workMode",
          "workingHours",
          "languages",
          "dealBreakers",
          "source",
          "updatedAt"
        ]
      },
      "CvImportSession": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "profileId": { "type": "string" },
          "profile": { "$ref": "#/components/schemas/Profile" },
          "rawText": { "type": "string" },
          "parsedFields": { "type": "object" },
          "lowConfidenceFields": { "type": "array", "items": { "type": "string" } },
          "status": { "enum": ["parsed", "confirmed", "discarded"], "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "confirmedAt": { "format": "date-time", "type": "string", "nullable": true }
        },
        "required": [
          "id",
          "profileId",
          "profile",
          "rawText",
          "parsedFields",
          "lowConfidenceFields",
          "status",
          "createdAt",
          "confirmedAt"
        ]
      },
      "ProfileMetadata": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "profileId": { "type": "string" },
          "profile": { "$ref": "#/components/schemas/Profile" },
          "usedFor": { "type": "object" },
          "lastAiAnalysis": { "format": "date-time", "type": "string", "nullable": true },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": ["id", "profileId", "profile", "usedFor", "lastAiAnalysis", "updatedAt"]
      },
      "ProfileInsight": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "profileId": { "type": "string" },
          "profile": { "$ref": "#/components/schemas/Profile" },
          "trigger": {
            "enum": [
              "cv_rejected",
              "job_saved",
              "interview_failed",
              "profile_incomplete",
              "skill_gap_detected"
            ],
            "type": "string"
          },
          "insight": { "type": "string" },
          "suggestedAction": { "type": "string" },
          "relatedProfileFields": { "type": "array", "items": { "type": "string" } },
          "isRead": { "type": "boolean" },
          "isActioned": { "type": "boolean" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "profileId",
          "profile",
          "trigger",
          "insight",
          "suggestedAction",
          "relatedProfileFields",
          "isRead",
          "isActioned",
          "createdAt"
        ]
      },
      "Profile": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "fullName": { "type": "string" },
          "phone": { "type": "string" },
          "address": { "type": "string" },
          "currentRole": { "type": "string" },
          "seniorityLevel": {
            "nullable": true,
            "enum": ["entry", "junior", "mid", "senior", "lead", "manager"],
            "type": "string"
          },
          "yearsOfExperience": { "type": "number", "nullable": true },
          "location": { "type": "string", "nullable": true },
          "workPreference": {
            "nullable": true,
            "enum": ["remote", "onsite", "hybrid", "flexible"],
            "type": "string"
          },
          "source": { "enum": ["user", "cv_parse", "ai_suggest"], "type": "string" },
          "confidence": { "type": "number" },
          "linkedin": { "type": "string" },
          "portfolio": { "type": "string" },
          "cvUrl": { "type": "string", "nullable": true },
          "cvFileName": { "type": "string", "nullable": true },
          "cvS3Key": { "type": "string", "nullable": true },
          "isPublic": { "type": "boolean" },
          "visibilitySettings": { "type": "object" },
          "profileSkills": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProfileSkill" }
          },
          "profileExperiences": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProfileExperience" }
          },
          "profileProjects": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProfileProject" }
          },
          "careerIntent": { "$ref": "#/components/schemas/CareerIntent" },
          "workPreferences": { "$ref": "#/components/schemas/WorkPreferences" },
          "cvImportSessions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CvImportSession" }
          },
          "metadata": { "$ref": "#/components/schemas/ProfileMetadata" },
          "insights": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProfileInsight" }
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "skills": { "type": "array", "items": { "type": "string" } },
          "experience": { "type": "array", "items": { "type": "object" } },
          "education": { "type": "array", "items": { "type": "object" } },
          "preferredIndustries": { "type": "array", "items": { "type": "string" } },
          "preferredJobTypes": { "type": "array", "items": { "type": "string" } },
          "preferredLocations": { "type": "array", "items": { "type": "string" } },
          "minSalaryExpectation": { "type": "number" },
          "completenessScore": { "type": "number" }
        },
        "required": [
          "id",
          "userId",
          "user",
          "fullName",
          "phone",
          "address",
          "currentRole",
          "seniorityLevel",
          "yearsOfExperience",
          "location",
          "workPreference",
          "source",
          "confidence",
          "linkedin",
          "portfolio",
          "cvUrl",
          "cvFileName",
          "cvS3Key",
          "isPublic",
          "visibilitySettings",
          "profileSkills",
          "profileExperiences",
          "profileProjects",
          "careerIntent",
          "workPreferences",
          "cvImportSessions",
          "metadata",
          "insights",
          "createdAt",
          "updatedAt",
          "skills",
          "experience",
          "education",
          "preferredIndustries",
          "preferredJobTypes",
          "preferredLocations",
          "minSalaryExpectation",
          "completenessScore"
        ]
      },
...
      "CV": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "jobId": { "type": "string" },
          "job": { "$ref": "#/components/schemas/Job" },
          "name": { "type": "string" },
          "content": { "type": "object" },
          "template": { "type": "string" },
          "score": { "type": "number" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "userId",
          "user",
          "jobId",
          "job",
          "name",
          "content",
          "template",
          "score",
          "createdAt"
        ]
      },
      "InterviewSchedule": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "trackerId": { "type": "string" },
          "tracker": { "$ref": "#/components/schemas/JobTracker" },
          "roundName": { "type": "string" },
          "scheduledAt": { "format": "date-time", "type": "string" },
          "type": {
            "enum": ["technical", "hr", "cultural", "system_design", "coding", "other"],
            "type": "string"
          },
          "locationUrl": { "type": "string" },
          "notes": { "type": "string" },
          "prepTips": { "type": "object", "nullable": true },
          "googleEventId": { "type": "string", "nullable": true },
          "calendarSyncedAt": { "format": "date-time", "type": "string", "nullable": true },
          "durationMinutes": { "type": "number" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "trackerId",
          "tracker",
          "roundName",
          "scheduledAt",
          "type",
          "locationUrl",
          "notes",
          "prepTips",
          "googleEventId",
          "calendarSyncedAt",
          "durationMinutes",
          "createdAt",
          "updatedAt"
        ]
      },
      "JobTracker": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "jobId": { "type": "string" },
          "job": { "$ref": "#/components/schemas/Job" },
          "manualTitle": { "type": "string" },
          "manualCompany": { "type": "string" },
          "manualUrl": { "type": "string" },
          "status": {
            "enum": ["saved", "applied", "interview", "offer", "rejected"],
            "type": "string"
          },
          "appliedAt": { "format": "date-time", "type": "string" },
          "cvId": { "type": "string" },
          "cv": { "$ref": "#/components/schemas/CV" },
          "notes": { "type": "string" },
          "nextActionDate": { "format": "date-time", "type": "string" },
          "isReminderSent": { "type": "boolean" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "interviews": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/InterviewSchedule" }
          }
        },
        "required": [
          "id",
          "userId",
          "user",
          "jobId",
          "job",
          "manualTitle",
          "manualCompany",
          "manualUrl",
          "status",
          "appliedAt",
          "cvId",
          "cv",
          "notes",
          "nextActionDate",
          "isReminderSent",
          "createdAt",
          "updatedAt",
          "interviews"
        ]
      },
      "RefreshToken": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "tokenHash": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "isRevoked": { "type": "boolean" },
          "expiresAt": { "format": "date-time", "type": "string" },
          "replacedBy": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "tokenHash",
          "userId",
          "user",
          "isRevoked",
          "expiresAt",
          "replacedBy",
          "createdAt"
        ]
      },
      "CoverLetter": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "jobId": { "type": "string" },
          "job": { "$ref": "#/components/schemas/Job" },
          "content": { "type": "string" },
          "tone": { "type": "string" },
          "language": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "userId",
          "user",
          "jobId",
          "job",
          "content",
          "tone",
          "language",
          "createdAt"
        ]
      },
      "JobAlert": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "isActive": { "type": "boolean" },
          "frequency": { "enum": ["daily", "weekly"], "type": "string" },
          "channels": { "type": "array", "items": { "type": "string", "enum": ["email"] } },
          "lastSentAt": { "format": "date-time", "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "userId",
          "user",
          "isActive",
          "frequency",
          "channels",
          "lastSentAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserCredits": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "balance": { "type": "number" },
          "lastRefillDate": { "format": "date-time", "type": "string", "nullable": true },
          "updatedAt": { "format": "date-time", "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        },
        "required": ["id", "userId", "balance", "lastRefillDate", "updatedAt", "user"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "passwordHash": { "type": "string", "nullable": true },
          "isVerified": { "type": "boolean" },
          "isBanned": { "type": "boolean" },
          "verificationToken": { "type": "string", "nullable": true },
          "resetPasswordToken": { "type": "string", "nullable": true },
          "resetPasswordExpires": { "format": "date-time", "type": "string", "nullable": true },
          "role": { "enum": ["user", "admin"], "type": "string" },
          "googleId": { "type": "string", "nullable": true },
          "githubId": { "type": "string", "nullable": true },
          "appleId": { "type": "string", "nullable": true },
          "avatarUrl": { "type": "string", "nullable": true },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "profile": { "$ref": "#/components/schemas/Profile" },
          "trackers": { "type": "array", "items": { "$ref": "#/components/schemas/JobTracker" } },
          "cvs": { "type": "array", "items": { "$ref": "#/components/schemas/CV" } },
          "refreshTokens": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RefreshToken" }
          },
          "coverLetters": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CoverLetter" }
          },
          "jobAlert": { "$ref": "#/components/schemas/JobAlert" },
          "credits": { "$ref": "#/components/schemas/UserCredits" }
        },
        "required": [
          "id",
          "email",
          "passwordHash",
          "isVerified",
          "isBanned",
          "verificationToken",
          "resetPasswordToken",
          "resetPasswordExpires",
          "role",
          "googleId",
          "githubId",
          "appleId",
          "avatarUrl",
          "createdAt",
          "updatedAt",
          "profile",
          "trackers",
          "cvs",
          "refreshTokens",
          "coverLetters",
          "jobAlert",
          "credits"
        ]
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "title": { "type": "string" },
          "message": { "type": "string" },
          "type": { "enum": ["info", "success", "warning", "error"], "type": "string" },
          "read": { "type": "boolean" },
          "link": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "userId",
          "user",
          "title",
          "message",
          "type",
          "read",
          "link",
          "createdAt"
        ]
      }
    }
  }
}